+++ GP/Gp.cpp
@@ -304,7 +304,7 @@ int CGpApp::ExitInstance()
 BOOL CGpApp::PreTranslateMessage(MSG *pMsg)
 {
     if (pMsg->message == WM_CHAR && pMsg->wParam == ' ')
-        m_pMainWnd->PostMessage(WM_COMMAND, MAKEWPARAM(ID_PBCK_NEXT, 0));
+        m_pMainWnd->PostMessage(WM_COMMAND, MAKEWPARAM(uint16_t(ID_PBCK_NEXT), uint16_t(0)));
     return CWinAppEx::PreTranslateMessage(pMsg);
 }
